rename linear op, flow, and transform to copy. move node to uniflow root directory
