+++ tests/test_node.py
@@ -1,5 +1,5 @@
 import unittest
-from uniflow.node.node import Node
+from uniflow.node import Node
 
 
 class TestNode(unittest.TestCase):
