+++ uniflow/flow/extract/extract_pdf_flow.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Sequence
 
 from uniflow.constants import EXTRACT
 from uniflow.flow.flow import Flow
-from uniflow.node.node import Node
+from uniflow.node import Node
 from uniflow.op.extract.load.pdf_op import ProcessPDFOp
 from uniflow.op.model.model_op import LLMDataPreprocessor
