+++ uniflow/flow/transform/__init__.py
@@ -5,18 +5,22 @@
 # __init_subclass__
 
 
-from uniflow.flow.transform.transform_huggingface_flow import \
-    TransformHuggingFaceFlow  # noqa: F401, F403
-from uniflow.flow.transform.transform_linear_flow import \
-    TransformLinearFlow  # noqa: F401, F403
-from uniflow.flow.transform.transform_lmqg_flow import \
-    TransformLMQGFlow  # noqa: F401, F403
-from uniflow.flow.transform.transform_openai_flow import \
-    TransformOpenAIFlow  # noqa: F401, F403
+from uniflow.flow.transform.transform_copy_flow import (  # noqa: F401, F403
+    TransformCopyFlow,
+)
+from uniflow.flow.transform.transform_huggingface_flow import (  # noqa: F401, F403
+    TransformHuggingFaceFlow,
+)
+from uniflow.flow.transform.transform_lmqg_flow import (  # noqa: F401, F403
+    TransformLMQGFlow,
+)
+from uniflow.flow.transform.transform_openai_flow import (  # noqa: F401, F403
+    TransformOpenAIFlow,
+)
 
 __all__ = [
     "TransformOpenAIFlow",
     "TransformHuggingFaceFlow",
     "TransformLMQGFlow",
-    "TransformLinearFlow",
+    "TransformCopyFlow",
 ]
