+++ uniflow/flow/transform/transform_huggingface_flow.py
@@ -3,7 +3,7 @@ from typing import Any, Dict, Sequence
 
 from uniflow.constants import TRANSFORM
 from uniflow.flow.flow import Flow
-from uniflow.node.node import Node
+from uniflow.node import Node
 from uniflow.op.model.model_op import LLMDataProcessor, ModelOp
 from uniflow.op.prompt_schema import GuidedPrompt
